refactor: Update uint to uint256 for consistency across contracts and improve precision

+++ script/DeployFactory.s.sol
@@ -9,8 +9,8 @@ contract DeployFactory is Script {
     function run() external {
         // --- Configuration (override via env vars or CLI flags) ---
         address usdc = vm.envAddress("USDC_ADDRESS");
-        uint creationFee = vm.envOr("CREATION_FEE", uint(10e6)); // $10 USDC (6 decimals)
-        uint initialLiquidity = vm.envOr("INITIAL_LIQUIDITY", uint(5e6)); // $5 seeded into market
+        uint256 creationFee = vm.envOr("CREATION_FEE", uint256(10e6)); // $10 USDC (6 decimals)
+        uint256 initialLiquidity = vm.envOr("INITIAL_LIQUIDITY", uint256(5e6)); // $5 seeded into market
 
         uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
 
@@ -20,12 +20,8 @@ contract DeployFactory is Script {
         PredictionMarket implementation = new PredictionMarket();
 
         // 2. Deploy the factory, pointing it at the implementation
-        PredictionMarketFactory factory = new PredictionMarketFactory(
-            usdc,
-            address(implementation),
-            creationFee,
-            initialLiquidity
-        );
+        PredictionMarketFactory factory =
+            new PredictionMarketFactory(usdc, address(implementation), creationFee, initialLiquidity);
 
         console.log("=== BizFun Deployment ===");
         console.log("Implementation deployed at:", address(implementation));
